sentinel/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sentinel-scanner"
version = "1.0.0"
description = "DevSecOps secret and misconfiguration scanner for codebases"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Hafiz Saad Tanvir" }]
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
sentinel = "sentinel.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["sentinel*"]
